push weekly algo

+++ radix_sort/sort.h
@@ -4,7 +4,7 @@
 #include <math.h>
 
 void print_array(const int *array, size_t size);
-int my_pow(int base, int exponent);
+int my_pow(double base, int exponent);
 void radix_sort(int *array, size_t size);
 int numLength(int num);
 int getMax(int *array, int n);
